Add Parser class and UI class to clean up code with OOP concepts
