+++ data/blue.txt
@@ -1,6 +1,5 @@
-T | 1 | Prepare for interview
-T | 1 | Brush Teeth
-T | 0 | Tuition
-D | 0 | CS2105 Assignment | 2/12/2024 1800
-D | 0 | cs2100 tutorial | 23/10/2020 1330
+D | 1 | CS2105 Assignment | 2/12/2024 1800
 D | 0 | Submit homework | 2/12/2024 1730
+D | 0 | bring blue for a walk | 31/8/2024 1900
+T | 1 | Prepare for interview
+T | 0 | homework
